requirements.txt
absl-py==2.1.0
array-record==0.4.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
astunparse==1.6.3
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
cycler==0.12.1
debugpy @ file:///D:/bld/debugpy_1707444604970/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
dm-tree==0.1.8
etils==1.5.2
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1704921103267/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
flatbuffers==1.12
fsspec==2024.6.0
gast==0.4.0
google-auth==2.30.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
idna==3.7
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1710971335535/work
importlib_resources==6.4.0
ipykernel @ file:///D:/bld/ipykernel_1717717683217/work
ipython @ file:///D:/bld/ipython_1701831845989/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1716472197302/work
jupyter_core @ file:///D:/bld/jupyter_core_1710257377578/work
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.4.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
numpy==1.23.0
oauthlib==3.2.2
opt-einsum==3.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1718189413536/work
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.3.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1715777629804/work
promise==2.3
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1718047967974/work
protobuf==3.20.3
psutil @ file:///D:/bld/psutil_1705722541723/work
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
pyparsing==3.1.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
pywin32==306
pyzmq @ file:///D:/bld/pyzmq_1715024536945/work
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.0
tensorflow-datasets==4.9.2
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.15.0
termcolor==2.4.0
toml==0.10.2
tornado @ file:///D:/bld/tornado_1717722891796/work
tqdm==4.66.4
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
urllib3==2.2.2
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
Werkzeug==3.0.3
wrapt==1.16.0
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1718013267051/work
